mkm: Add test script for `cn test`, modify other files to make it work
